my_text_aggregator/requirements.txt
PyPDF2
pdfplumber
pdf2image
pytesseract
python-docx
odfpy
